+++ Code/sources/DataInterface.cpp
@@ -5,7 +5,7 @@
     @brief Returns the entire raw_sensor_data_ private member
     @return Vector of vector of pairs, i.e. a 2D matrix with complex numbers.
 */
-std::vector<std::vector<std::pair<int, int>>> DataInterface::get_raw_sensor_data(){
+std::vector<std::vector<std::pair<double, double>>> DataInterface::get_raw_sensor_data(){
     return raw_sensor_data_;
 }
 
@@ -15,7 +15,7 @@ std::vector<std::vector<std::pair<int, int>>> DataInterface::get_raw_sensor_data
     @param data_idx The data index from which to return the pair
     @return Pair of integers, i.e. a complex number.
 */
-std::pair<int, int> DataInterface::get_raw_sensor_data_point(int freq_idx, int data_idx){
+std::pair<double, double> DataInterface::get_raw_sensor_data_point(int freq_idx, int data_idx){
     return raw_sensor_data_.at(freq_idx).at(data_idx);
 }
 
@@ -23,7 +23,7 @@ std::pair<int, int> DataInterface::get_raw_sensor_data_point(int freq_idx, int d
     @brief Sets the entire raw_sensor_data_ private member
     @param data Vector of vector of pairs, i.e. a 2D matrix with complex numbers.
 */
-void DataInterface::set_raw_sensor_data(std::vector<std::vector<std::pair<int, int>>> data) {
+void DataInterface::set_raw_sensor_data(std::vector<std::vector<std::pair<double, double>>> data) {
         raw_sensor_data_ = data;
 }
 
@@ -34,7 +34,7 @@ void DataInterface::set_raw_sensor_data(std::vector<std::vector<std::pair<int, i
     @param freq_idx The frequency index at which to update the pair value
     @param data_idx The data index at which to update the pair value
 */
-void DataInterface::set_raw_sensor_data_point(int real, int img, int freq_idx, int data_idx) {
+void DataInterface::set_raw_sensor_data_point(double real, double img, int freq_idx, int data_idx) {
         raw_sensor_data_.at(freq_idx).at(data_idx) = {real, img};
 }
 
@@ -42,7 +42,7 @@ void DataInterface::set_raw_sensor_data_point(int real, int img, int freq_idx, i
     @brief Prints the data from the 2D matrix of complex numbers in a human-readable format
     @param data Vector of vector of pairs, i.e. a 2D matrix with complex numbers.
 */
-void DataInterface::print_data(std::vector<std::vector<std::pair<int, int>>> data){
+void DataInterface::print_data(std::vector<std::vector<std::pair<double, double>>> data){
     std::cout << "Data point\t";
     for (int i = 0; i < data.size(); i++) {
         std::cout << "Freq " << i + 1 << "\t";
